+++ src/main/java/org/example/goormssd/usermanagementbackend/controller/AuthController.java
@@ -17,7 +17,7 @@ import org.example.goormssd.usermanagementbackend.dto.response.FindEmailResponse
 import org.example.goormssd.usermanagementbackend.dto.response.LoginResponseDto;
 import org.example.goormssd.usermanagementbackend.dto.response.RefreshTokenDto;
 import org.example.goormssd.usermanagementbackend.service.AuthService;
-import org.example.goormssd.usermanagementbackend.service.dto.LoginResult;
+import org.example.goormssd.usermanagementbackend.dto.response.LoginResult;
 import org.example.goormssd.usermanagementbackend.config.security.JwtUtil;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
