+++ src/main/java/org/example/goormssd/usermanagementbackend/dto/response/LoginResult.java
@@ -1,8 +1,7 @@
-package org.example.goormssd.usermanagementbackend.service.dto;
+package org.example.goormssd.usermanagementbackend.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import org.example.goormssd.usermanagementbackend.dto.response.LoginUserDto;
 
 
 @Getter
